Reorganize: Action::Result to own file
